backend/requirements.txt
flask
flask_restful
flask_cors
gtts

sqlalchemy

mariadb
python-dotenv
mysql-connector-python
mysqlclient
flask-sqlalchemy
retry